feat: parameterized the properties panel subheading getter logic
Closes #1290
