+++ packages/form-js-viewer/src/render/components/Util.js
@@ -27,6 +27,21 @@ export function gridColumnClasses(formField) {
   );
 }
 
+export function textToLabel(text) {
+  if (typeof text != 'string') return null;
+
+  for (const line of text.split('\n')) {
+    const displayLine = line.trim();
+
+    // we use the first non-whitespace line in the text as label
+    if (displayLine !== '') {
+      return displayLine;
+    }
+  }
+
+  return null;
+}
+
 export function prefixId(id, formId, indexes) {
   let result = 'fjs-form';
