+++ packages/form-js-viewer/src/render/components/form-fields/Datetime.js
@@ -224,4 +224,7 @@ Datetime.config = {
 
     return { ...defaults, ...options };
   },
+  getSubheading: (field) => {
+    return field.dateLabel || field.timeLabel;
+  },
 };
