+++ packages/form-js-viewer/src/render/components/form-fields/Default.js
@@ -19,4 +19,5 @@ Default.config = {
     components: [],
     ...options,
   }),
+  getSubheading: (field) => field.id,
 };
