+++ packages/form-js-viewer/src/render/components/form-fields/Image.js
@@ -50,4 +50,5 @@ Image.config = {
   name: 'Image view',
   group: 'presentation',
   create: (options = {}) => options,
+  getSubheading: (field) => field.alt,
 };
